requirements.txt
streamlit==1.31.0
anthropic==0.18.0
pyyaml==6.0.1
requests==2.31.0
jsonschema==4.21.1
python-dotenv==1.0.0
openai == 2.17.0
google-generativeai == 0.8.6